Fix printing bug for Mutex operator.
When printing with verbose output the location of the lock and unlock
calls in the original C file is printed if debug information is present.

Previously, the debug information for the lock call was used for both
the unlock and lock call. This resulted in the output saying that the
lock and unlock call were on the same line.